(QA-2297) BeakerTask: should inherit from Rake  TaskLib, and execute beaker
Make the BeakerTask class actually work.  more or less.  we still have
problems with a probable bug in beaker which prevents multiple flags on
the CLI.  We also change to naming it AcceptanceTask to make it uniform
across any underlying tooling. We should probably abstract this a bit
better, later, by registering a framework with RototillerTask.

We also add a bunch of testing for color_text and AcceptanceTask.
AcceptanceTask needs a bunch more testing, but we're up to 100%
coverage.  oh yeah, and we enabled simplecov for unit tests.