(QA-2445) Add support for setting boolean flags with rototiller
We currently do not have a way to turn on and off switches, aka:
boolean-flags/options. This change adds a switch parameter to add_flag
to specify is_boolean.  The override_env now operates on the option
name, and the default value can be used to default the switch to 'off'.

This is bad and we should feel bad. the API if very warty and needs
refactoring. (post v0.1.0)