(QA-2637) command #add_env
* add #add_env to command
* it will overwrite command name with its value
* FIXME: we mainly duplicate add_env here from rototiller_task, it can
  be put in a module
* add #to_str for EnvVar so it can be printed and find its value
* add #last to env_collection so one can easily find the last known
value of a collection of EnvVars.
* block_handling needed empty methods defined for nested blocks and
hashes. When a Command is created and has a nested add_env, for
instance, the add_env must be both a setter and a method (getter), even
though it is not actually called.
* ensure we return the EnvVar or Command when using #add_env or
 #add_command
* EnvVar (and other params) need their attribute names to match their
allowed attributes because we now must yield using EnvVar objects, so
the blocks populate the actual EnvVar